refactor(three-renderer): extract shared glyph entity base for MTEXT and SHAPE
Move duplicated mtext-renderer integration (draw, raycast, batching, selection box) into AcTrGlyphEntity so AcTrMText and AcTrShape only supply entity-specific render input.
