+++ packages/three-renderer/src/object/index.ts
@@ -1,4 +1,5 @@
 export * from './AcTrEntity'
+export * from './AcTrGlyphEntity'
 export * from './AcTrGroup'
 export * from './AcTrImage'
 export * from './AcTrLine'
